+++ codeclassroom/settings.py
@@ -25,6 +25,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rest_framework',
+    'rest_framework.authtoken',
+    'drf_yasg',
     'corsheaders',
     'app',
     'api',
@@ -67,11 +69,26 @@ WSGI_APPLICATION = 'codeclassroom.wsgi.application'
 REST_FRAMEWORK = {
     'DEFAULT_AUTHENTICATION_CLASSES': [
         'rest_framework.authentication.SessionAuthentication',
-        'rest_framework.authentication.BasicAuthentication'
+        'rest_framework.authentication.BasicAuthentication',
+        'rest_framework.authentication.TokenAuthentication',
     ],
     'DEFAULT_PERMISSION_CLASSES': [
         'rest_framework.permissions.IsAuthenticated',
     ]
+}
+
+# drf-yasg settings
+
+SWAGGER_SETTINGS = {
+    'SECURITY_DEFINITIONS': {
+         'DRF Token': {
+           'type': 'apiKey',
+           'name': 'Authorization',
+           'in': 'header'
+         }
+    },
+    'LOGIN_URL': 'login',
+    'LOGOUT_URL': 'logout',
 }
 
 # Database